+++ src/repo/project.ts
@@ -1,0 +1,7 @@
+import { repository } from '@chevre/domain';
+
+/**
+ * プロジェクトリポジトリ
+ */
+export class MongoRepository extends repository.Project {
+}